requirements.txt
aiogram==3.11.0
SQLAlchemy==2.0.31
APScheduler==3.10.4
pandas==2.2.2
python-dotenv==1.0.1
pytz==2024.1
cachetools==5.4.0
requests==2.32.3
aiogram_i18n==1.4
aiosqlite==0.20.0
matplotlib==3.9.1.post1
openpyxl==3.1.5
fluent.runtime==0.4.0